2018/1/7

+++ simulation/stdafx.cpp
@@ -12,6 +12,6 @@ std::map<int, database> data;
 
 fever_t fever_struct;
 
-double size;
+float size;
 double airtem;
 double tem_step;
